+++ DiaryApp/DiaryApp/Scene/Signin/View/ViewController/SignInViewController.swift
@@ -8,9 +8,8 @@
 import UIKit
 
 class SignInViewController: UIViewController {
-	
-	// window 객체 사용하기 위해서 싱글톤패턴으로 sceneDelegate 생성
-	let sceneDelegate = UIApplication.shared.connectedScenes.first!.delegate as! SceneDelegate
+	// 싱글톤 가져옴
+	let signInService = SignInService.shared
 	
 	@IBOutlet var signInButtons: [UIButton]!
 	@IBOutlet weak var titleLabel: UILabel!
@@ -38,44 +37,23 @@ class SignInViewController: UIViewController {
 		titleLabel.textColor = UIColor(hexString: "#7E76DE")
 	}
 	
-	// 로그인 로직
-	func getSignIn() {
-		// userDefaults 에 로그인 식별하기 위한 키값 저장
-		// 앱을 껐다켜도 로그인 유지
-		UserDefaults.standard.setValue(true, forKey: "authVerificationID")
-		UserDefaults.standard.synchronize()
-		
-		let storyboard = UIStoryboard(name: "SignIn", bundle: nil)
-		let homeViewController = storyboard.instantiateViewController(withIdentifier: "Home")
-		let homeNavigationController = UINavigationController(rootViewController: homeViewController)
-		
-		guard let window = sceneDelegate.window else { return }
-		
-		// 로그인 완료시 루트뷰 바꿔줌
-		window.rootViewController = homeNavigationController
-	}
-	
 	@IBAction func appleButtonTapped(_ sender: UIButton) {
-		getSignIn()
-		
-		print("애플 로그인함: ", UserDefaults.standard.value(forKey: "authVerificationID")!)
+		signInService.getSignIn()
+		goHomeVC()
 	}
 	
 	@IBAction func kakaoButtonTapped(_ sender: UIButton) {
-		getSignIn()
-		
-		print("카카오 로그인함: ", UserDefaults.standard.value(forKey: "authVerificationID")!)
+		signInService.getSignIn()
+		goHomeVC()
 	}
 	
 	@IBAction func googleButtonTapped(_ sender: UIButton) {
-		getSignIn()
-		
-		print("구글 로그인함: ", UserDefaults.standard.value(forKey: "authVerificationID")!)
+		signInService.getSignIn()
+		goHomeVC()
 	}
 	
 	@IBAction func naverButtonTapped(_ sender: UIButton) {
-		getSignIn()
-		
-		print("네이버 로그인함: ", UserDefaults.standard.value(forKey: "authVerificationID")!)
+		signInService.getSignIn()
+		goHomeVC()
 	}
 }
